+++ Form1.cs
@@ -13,9 +13,9 @@ namespace Spectres
     public partial class Form1 : Form
     {
         private MathCad mathCad;
-        int[] M = new int[3];
+        double[] M = new double[3];
         int N = 0;
-        int m = 0;
+        double m = 0;
 
         public Form1()
         {
@@ -33,7 +33,7 @@ namespace Spectres
         private void btnRazdel2_Click(object sender, EventArgs e)
         {
             if (!parametersAreCorrect()) return;
-            var formRazdel2 = new FormRazdel2(mathCad, new int[] { m }, N);
+            var formRazdel2 = new FormRazdel2(mathCad, new double[] { m }, N);
             formRazdel2.Show();
         }
 
@@ -41,7 +41,7 @@ namespace Spectres
         {
             if (
                 tb1M1.Text == tb1M2.Text || tb1M1.Text == tb1M3.Text || tb1M2.Text == tb1M3.Text
-                || !int.TryParse(tb1M1.Text, out M[0]) || !int.TryParse(tb1M2.Text, out M[1]) || !int.TryParse(tb1M3.Text, out M[2]) || !int.TryParse(tb2M.Text, out m) || !int.TryParse(tb2N.Text, out N)
+                || !double.TryParse(tb1M1.Text, out M[0]) || !double.TryParse(tb1M2.Text, out M[1]) || !double.TryParse(tb1M3.Text, out M[2]) || !double.TryParse(tb2M.Text, out m) || !int.TryParse(tb2N.Text, out N)
                 )
             {
                 MessageBox.Show("Проверьте введённые данные");
